Porting in uno's camera fade fix
Referencing https://github.com/uno1982 's pull request but re-creating it to fit the plugin better. Allows overriding the PendingPlayerController with a custom LocalPlayer class, also adds a base player controller class for the VRPlayerController or as standalone that implements a OnCameraManagerCreated event. This allows us to in BP specifically fade the screen when the camera manager is created, avoiding the flickering and level view until the player character is actually spawned and active. Being an implementable event lets the end user use platform specific fade nodes (like oculus's one) instead of the camera managers fade commands as well.