Added bIsPendingKill to grips
Lets us filter on simulated clients if we are running operations on the grip array inside of a post drop event. On simulated clients its not safe to just remove the grip from the array so it was paused until replication caught up originally. Now we also mark it as pending kill so that its entirely ignored and so users can filter against it.