OpenXRExpansionPlugin now compiling on Mac
We don't need to actually run OpenXR on a Mac but if we're targeting another platform (Android for the Quest for example) then we need to #ifdef out some OpenXR dependencies to let the plugin successfully compile. The plugin should compile with proper support for any other platform that supports OpenXR (apart from Mac).